requirements.txt
python-telegram-bot[job-queue,socks]==20.7
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pyinstaller==6.3.0
httpx[socks]==0.27.0
socksio==1.0.0

